Restructure of Duration Timer, No negative values in rest timer
