+++ FitnessApp/FitnessApp/Extensions/Extensions.swift
@@ -14,7 +14,10 @@ extension Double {
     func timeString() -> (hours: String, minutes: String, seconds: String) {
         let hours   = Int(self) / 3600
         let minutes = Int(self) / 60 % 60
-        let seconds = Int(self) % 60
+        var seconds = Int(self) % 60
+        if (seconds < 0) {
+            seconds = 0
+        }
         return (String(format:"%02i:%02i:%02i", hours, minutes, seconds),
                 String(format:"%02i:%02i", hours, minutes),
                 String(format:"%02i:%02i", minutes, seconds))
